Promote UnsafeCursor to multiplatform
I want to use this in okio-files.
